rename: 폴더 위치 변경
- 프로필 페이지에서 공통으로 사용되기 때문에 위치를 변경했습니다.
