+++ src/pages/Profile/components/CoverImage.tsx
